Do not fall for partial negative lookahead matches
Fix #65

A pattern like '*.!(x).!(y)' should not match a name like 'a.xyz.yab'.

In order to accomplish this, the lookahead must look ALL the way ahead,
even to the very end of the pattern.

After parsing the entire pattern, copy everything that comes after the
pattern into the negative lookahead section.