filter out consecutive ** more efficiently
Previously, consecutive globstars were being properly stripped out of
makeRe, but leaving them in there in the normal minimatch set
geometrically increases the amount of work that has to be done when glob
gets a pattern like 'a/**/**/b'.

Better to strip them out from the initial pattern, and never have to
worry about it again.