optimize common patterns, correct matching of .*
Previously, `.*` would match `.` and `..` in some cases where it
shouldn't, unlike the Bash 5 behavior targeted.

Also, this adds fast-path optimizations for:
- `*`
- `*.<some extension>`
- `.*`

which are by far the most common patterns in use, and ironically also
some of the least-performant regular expressions generated by this
library.

While the regular expression generated by makeRE will always have to
take the slower approach (lots of lookahead/lookbehind tests,
conditional anchors, etc.), in the case of `minimatch()` or
`Minimatch.test()`, or manually walking the MMRegExp objects in the set,
it will be *much* faster.