Prevent dots only at the start of repeat patterns
A pattern like `+(?)` will note that the `?` matches the start of the
pattern, and prevent it from matching a dot. However, this is not
"anything other than a dot, repeating", but rather "repeating, where the
first repetition doesn't start with a dot".

With this change, repetitive extglob patterns in the start position are
expanded such that the first instance of the pattern may not start with
a dot, but any subsequent repetitions may begin with a dot.

Fix: https://github.com/isaacs/minimatch/issues/211